Adding aurelia dialog, ui validations, controls and services calls, fixes over api departmenType

+++ Hahn.ApplicationProcess.February2021.Domain/Models/Examples/AssetCreationResponseExample.cs
@@ -16,7 +16,7 @@ namespace Hahn.ApplicationProcess.February2021.Domain.Models.Examples
                     ID = 6,
                     AssetName = "Asset100",
                     CountryOfDepartment = "Colombia",
-                    Deparment = Enums.DepartmentType.HQ,
+                    Deparment = (int)Enums.DepartmentType.HQ,
                     EMailAdressOfDepartment = "Asset100@asset.com",
                     IsBroken = false,
                     PurchaseDate = DateTimeOffset.Parse("01/01/2021")
